packagekit: move watchRedHatSubscription to packagekit page
This is only used in the packagekit page and doesn't make sense to be
part of a shared "packagekit" API as it uses a different dbus interface.
