+++ pkg/lib/packagekit.js
@@ -284,53 +284,6 @@ export function cancellableTransaction(method, arglist, progress_cb, signalHandl
     });
 }
 
-/**
- * Check Red Hat subscription-manager if if this is an unregistered RHEL
- * system. If subscription-manager is not installed or required (not a
- * Red Hat product), nothing happens.
- *
- * callback: Called with a boolean (true: registered, false: not registered)
- *           after querying subscription-manager once, and whenever the value
- *           changes.
- */
-export function watchRedHatSubscription(callback) {
-    const sm = cockpit.dbus("com.redhat.RHSM1");
-
-    function check() {
-        sm.call(
-            "/com/redhat/RHSM1/Entitlement", "com.redhat.RHSM1.Entitlement", "GetStatus", ["", ""])
-                .then(result => {
-                    const status = JSON.parse(result[0]);
-                    callback(status.valid);
-                })
-                .catch(ex => console.warn("Failed to query RHEL subscription status:", JSON.stringify(ex)));
-    }
-
-    // check if subscription is required on this system, i.e. whether there are any installed products
-    sm.call("/com/redhat/RHSM1/Products", "com.redhat.RHSM1.Products", "ListInstalledProducts", ["", {}, ""])
-            .then(result => {
-                const products = JSON.parse(result[0]);
-                if (products.length === 0)
-                    return;
-
-                // check if this is an unregistered RHEL system
-                sm.subscribe(
-                    {
-                        path: "/com/redhat/RHSM1/Entitlement",
-                        interface: "com.redhat.RHSM1.Entitlement",
-                        member: "EntitlementChanged"
-                    },
-                    () => check()
-                );
-
-                check();
-            })
-            .catch(ex => {
-                if (ex.problem != "not-found")
-                    console.warn("Failed to query RHSM products:", JSON.stringify(ex));
-            });
-}
-
 /* Support for installing missing packages.
  *
  * First call check_missing_packages to determine whether something
